Extractor: move parse_* functions to separate file
