+++ src/priviblur_extractor/parse/__init__.py
@@ -1,1 +1,1 @@
-from .parsers import parse_timeline, parse_blog_timeline, parse_post_list
+from .base import parse_timeline, parse_blog_timeline, parse_post_list
